Redirect to the `fine-print/` directory
Rather than relying on Nginx to realize it's a dir, go to the dir directly. The response from the service in the previous iteration would be 301 Permanently Moved, which isn't so great. It generates a second request every time someone clicks this link. :frowning:

Noticed this in the test output for https://github.com/benbalter/benbalter.github.com/pull/211