Revised
Hi,

Amazing site! I got directed here yesterday by a very nice developer, who recommended you as one of the go-to people for Jekyll, and thought there was something to be added to the article; and a few areas I think readability can be improved, as well as the understanding of collections.

I'd love to get feedback, and of course give something back.

Hope you are having a great fathers day!