Fixed table formatting
I was looking for the reference / [link to this post](http://ben.balter.com/2015/11/23/why-open-source/) and noticed when I did that the table was mis-formatted... Maybe this is a Jekyll 2 -> 3 issue?

Anyways, I fixed the table rendering (looks ok in GitHub browser preview) to correct the issue. It looks like GitHub markdown previews don't support the `{: .table style="..."}` formatting though, so I'm not 100% that will look correct after the update. It renders incorrectly in [GitHub](https://github.com/benbalter/benbalter.github.com/blob/master/_posts/2015-11-23-why-open-source.md), but maybe it'll be fine when Pages / Jekyll renders it?